Move the MARC models into a dedicated file
