+++ packages/shared/sierra-client/src/index.ts
@@ -2,5 +2,6 @@ import SierraClient from './SierraClient';
 import HttpSierraClient from './HttpSierraClient';
 import MockSierraClient from './MockSierraClient';
 
-export { PatronRecord, Role, varFieldTags } from './patron';
+export { PatronRecord, Role } from './patron';
+export { varFieldTags } from './marc';
 export { HttpSierraClient, MockSierraClient, SierraClient };
